safari-600.7-branch/LayoutTests/platform/gtk/accessibility/aria-slider-required-attributes-expected.txt

This tests the exposure of attributes for an ARIA slider.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS axSlider.role is 'AXRole: AXSlider'
PASS axSlider.intValue is 50
PASS axSlider.minValue is 0
PASS axSlider.maxValue is 100
PASS axSlider.intValue is 75
PASS axSlider.intValue is 35
PASS valueChangedNotifications is 2
PASS successfullyParsed is true

TEST COMPLETE
